+++ Structurizr.Core/Model/Component.cs
@@ -63,7 +63,7 @@ namespace Structurizr
         {
             get
             {
-                return Parent.CanonicalName + CanonicalNameSeparator + FormatForCanonicalName(Name);
+                return new CanonicalNameGenerator().Generate(this);
             }
         }