+++ Structurizr.Core/Model/Person.cs
@@ -23,7 +23,7 @@ namespace Structurizr
         {
             get
             {
-                return CanonicalNameSeparator + FormatForCanonicalName(Name);
+                return new CanonicalNameGenerator().Generate(this);
             }
         }